% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjSameName.R
\name{adj_rename}
\alias{adj_rename}
\title{adj_rename}
\usage{
adj_rename(adjL)
}
\arguments{
\item{adjL}{list of adjacency matrices}
}
\value{
a list of adjacency matrices with the same rows and columns name.
}
\description{
This function constructs a list of adjacency matrices with the
same row and column names for all the matrices. The output is the object
needed for \code{\link{consensusNet}} function.
}
\examples{
data("tryL_data")
adj_rename(tryL_data)
}
