% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjL_data}
\alias{adjL_data}
\title{Adjacency Data}
\format{
## `adjL_data`
A list of 2 objects:
\describe{
  \item{AdjMatrix1}{Adjacency matrix;}
  \item{AdjMatrix2}{Adjacency matrix.}
}
}
\usage{
adjL_data
}
\description{
List of 2 adjacency matrices data type.
}
\keyword{datasets}
