% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityWeight.R
\name{densityNet}
\alias{densityNet}
\title{densityNet}
\usage{
densityNet(graphL)
}
\arguments{
\item{graphL}{the list of weighted graphs in igraph format.}
}
\value{
the quantile of the mean density distribution,
the quantile of the mean density distribution without the zeros,
plot density distribution without the zeros
}
\description{
This function creates a density plot of the different graphs mean
weights. It can be used to search the final Threshold for the Consensus
Network starting from similar networks.
}
\examples{
data("graphL_data")
densityNet(graphL_data)
}
