% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleL_data}
\alias{exampleL_data}
\title{Example Data}
\format{
## `exampleL_data`
A list of 3 objects:
\describe{
\item{Gene_Expression}{subset of Gene expression data;}
  \item{Methy_Expression}{subset of Methylation data;}
  \item{Mirna_Expression}{subset of Mirna data.}
}
}
\source{
<https://portal.gdc.cancer.gov/>
}
\usage{
exampleL_data
}
\description{
3 data types: Gene_Expression, Methy_Expression and Mirna_Expression data
from patients with Glioblastoma
}
\keyword{datasets}
