% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunction.R
\name{get_lower_tri_noDiag}
\alias{get_lower_tri_noDiag}
\title{get_lower_tri_noDiag}
\usage{
get_lower_tri_noDiag(cormat)
}
\arguments{
\item{cormat}{matrix}
}
\description{
create a lower triangle of the matrix without the diagonal
}
\keyword{internal}
