% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graphL_data}
\alias{graphL_data}
\title{Graph Data}
\format{
## `graphL_data`
A list of 2 objects:
\describe{
  \item{Graph1}{Graph firt layer;}
  \item{Graph2}{Graph second layer.}
}
}
\usage{
graphL_data
}
\description{
List of 2 graphs of igraph class type.
}
\keyword{datasets}
