% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardWeightedMatrix.R
\name{JWmatrix}
\alias{JWmatrix}
\title{JWmatrix}
\usage{
JWmatrix(graphL)
}
\arguments{
\item{graphL}{list of graphs as igraph objects with the same nodes.}
}
\value{
weighted Jaccard distance matrix
}
\description{
This function computes the Jaccard weighted matrix distance
between all the pairs of graphs.
}
\examples{
data("graphL_data")
JWmatrix(graphL_data)
}
