% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardWeightedMean.R
\name{JWmean}
\alias{JWmean}
\title{JWmean}
\usage{
JWmean(graphL)
}
\arguments{
\item{graphL}{list of different graphs in igraph format with same nodes.}
}
\value{
a number: the mean distance
}
\description{
This function computes the Mean Weighted Jaccard Distance for
Multilayer Networks.
}
\examples{
data("graphL_data")
JWmean(graphL_data)
}
