% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkMeasures.R
\name{measuresNet}
\alias{measuresNet}
\title{measuresNet}
\usage{
measuresNet(graphL, nodes.measures = TRUE)
}
\arguments{
\item{graphL}{a list of graphs as igraphs objects.}

\item{nodes.measures}{logical, if falso it computes only graph measures, if
true it computes also nodes measures (default TRUE).}
}
\value{
list of measure for each layer.
}
\description{
This function computes graphs and nodes measures to analyse all
 the layers in one shot.
}
\examples{
data("graphL_data")
measuresNet(graphL_data)

}
