% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotC}
\alias{plotC}
\title{plotC}
\usage{
plotC(graph, ...)
}
\arguments{
\item{graph}{a graph}

\item{...}{other parameter}
}
\value{
plot
}
\description{
The function plots the network without isolated nodes.
}
\examples{
data("graphL_data")
plotC(graphL_data[[1]])
}
