% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecificFunction.R
\name{specificNet}
\alias{specificNet}
\title{specificNet}
\usage{
specificNet(graphL, graph.consensus)
}
\arguments{
\item{graphL}{a list of graphs as igraphs objects.}

\item{graph.consensus}{graphConsensus output of the
\code{\link{consensusNet}} function.}
}
\value{
Case Specific Networks one for each layer and percentage of
specificity.
}
\description{
The function creates Case Specific Networks one for each layer
to give information of the peculiar layer not present in the Consensus.
}
\examples{
data("graphL_data")
data("adjL_data")
myConsensus <- consensusNet(adjL_data)
specificNet(graphL_data, myConsensus$graphConsensus)
}
