% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdConsensus.R
\name{thresholdNet}
\alias{thresholdNet}
\title{thresholdNet}
\usage{
thresholdNet(sim.graphL, threshold = 0.5)
}
\arguments{
\item{sim.graphL}{a list of similarGraphs output of the
\code{\link{consensusNet}} function.}

\item{threshold}{different threshold to compute.}
}
\value{
a new consensus network igraph object.
}
\description{
The function reconstructs the Consensus Network with different
thresholding after the \code{\link{consensusNet}} function starting from
similar graphs.
}
\examples{
data("adjL_data")
myConsensus <- consensusNet(adjL_data)
thresholdNet(myConsensus$similarGraphs)
}
