% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tryL_data}
\alias{tryL_data}
\title{try Data}
\format{
## `tryL_data`
A list of 2 objects:
\describe{
  \item{AdjMatrix1}{Adjacency matrix;}
  \item{AdjMatrix2}{Adjacency matrix.}
}
}
\usage{
tryL_data
}
\description{
Random data with different nodes name in a list of 2 adjacency matrices.
}
\keyword{datasets}
